+++ lib/utils/functional/functional.dart
@@ -1,2 +1,4 @@
 library geoff.utils.functional;
 
+
+export './fp.dart';