+++ lib/utils/utils.dart
@@ -5,3 +5,4 @@ export './system/system.dart';
 export './maths/maths.dart';
 export './alarm.dart';
 export './functional/functional.dart';
+export './time/time.dart';