+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## **Version 0.2**
 ### **0.2.5**
+- Changed a bunch of classes with only `static` methods to be `abstract` and added private constructors to prevent extending
+- Added the `functional` library, which contains a few funtional programming methods
+
 Finally adding stuff to the `models` package this update!
 - Added `GeoffUser` and `KeycloakUser`
 - Added `ModelUtils`