+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### **0.2.5**
 - Changed a bunch of classes with only `static` methods to be `abstract` and added private constructors to prevent extending
 - Added the `functional` library, which contains a few funtional programming methods.
+- Added `time` library with some utc helpers.
 
 Finally adding stuff to the `models` package this update!
 - Added `GeoffUser` and `KeycloakUser`